SOLR-16165: Rare deadlock in SlotAcc initialization (#819)
